[#187]♻️refactor : compress / fileUpload 로직 분리
